Adds SearchView on SearchFragment.kt

+++ app/src/main/java/cl/maleb/mercadolibre/challenge/api/list/ReferencePrice.kt
@@ -1,7 +1,7 @@
 package cl.maleb.mercadolibre.challenge.api.list
 
 data class ReferencePrice(
-    val amount: Int? = null,
+    val amount: Double? = null,
     val conditions: PriceConditions? = null,
     val currency_id: String? = null,
     val exchange_rate_context: String? = null,
